Prototype a junit syntax
Inherently less powerful, because we can't generate tests
for several language versions
